test(ar): collect release target manifest checks
